Port existing .c and .h files of vi to Linux
[git-p4: depot-paths = "//depot/openwatcom/": change = 23909]

+++ bld/vi/h/rtns2.h
@@ -34,7 +34,7 @@
 void PositionScrollThumb( void );
 
 /* biosos2.c */
-void MyVioShowBuf( unsigned short, unsigned short );
+void MyVioShowBuf( unsigned, int );
 
 /* biosqnx.c */
 int KeyboardInit( void );
